Rename system_ plugin
