+++ src/middlewared/middlewared/plugins/system/coredump.py
